refactor(std): move the dimension-order helpers to burn-std
`dim_order`, `nested_dim_order` and `is_contiguous_order` answer a question
about a tensor's strides alone — which dimension is innermost, which next,
whether the layout is dense or merely nested — and nothing in them is about
fusion. They lived in `burn-cubecl-fusion`'s launch planner because that was
the first place to ask. The elementwise kernels in `burn-cubecl` are about to
ask the same question, and `burn-cubecl` cannot depend on the fusion crate.

Both crates already depend on `burn-std`, which is where `is_dense` and
`is_contiguous` sit, so the helpers and their tests move there as
`burn_std::tensor::layout`. The fusion planner re-exports them from the same
path it always did; `strides_for` and `permuted_innermost_axis` stay, being
about the planner's own reference layout.

No behaviour changes.
